+++ src/js/Etoile/service/index.js
@@ -1,0 +1,2 @@
+export { default as CategoryService } from './Category/index.js';
+export { default as MenuService } from './Menu/index.js';
